requirements.txt
bumpversion
requests
ipython
colorama
python-baseconv


